requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.2
streamlit==1.28.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
pytest==7.4.3
black==23.11.0